requirements.txt
python = 3.7.16
pytorch = 1.8
transformers = 4.23.1
numpy=1.21.2
fire=0.4.0
jieba_fast=0.39
csv=1.0
scikit-learn=1.1.2
